requirements.txt
asgiref==3.4.1
Django==3.2.7
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-humanize==0.1.2
django-js-asset==1.2.2
humanize==3.11.0
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.2
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.6
